feat(deps): update to typeorm 0.3.4
BREAKING CHANGE: Updates to TypeORM 0.3.4.

Connections are gone in TypeORM 0.3.x as is the connection manager.
Things now revolve around exported DataSource. Therefore, the
TypeORMCoreModule has to maintain its own connection manager for the
purposes of `keepConnectionAlive`

Otherwise, the changes are mostly going from `createConnection()` to
`new DataSource()`

For the sake of consistency, `keepConnectionAlive` was renamed to
`keepDataSourceAlive` which is a breaking change.